+++ src/main/java/net/instantgratification/mcainclusive/ducks/GeneticsDuck.java
@@ -1,0 +1,9 @@
+// Copyright (C) 2026 Dasik (Rifaditya) | GNU GPLv3
+package net.instantgratification.mcainclusive.ducks;
+
+public interface GeneticsDuck {
+    float getLeftBreastSize();
+    void setLeftBreastSize(float val);
+    float getRightBreastSize();
+    void setRightBreastSize(float val);
+}